+++ src/features/team/components/team-columns.tsx
@@ -10,7 +10,7 @@ import {
 import { MoreHorizontal } from "lucide-react";
 import { Button } from "~/_components/ui/button";
 import { type teamPlayers } from "~/utils/players";
-import { dbDropPlayer } from "~/features/team/database/teamActions";
+import { dbDropPlayer } from "~/features/team/actions/teamActions";
 import { toast } from "sonner";
  
 async function dropPlayer(player: teamPlayers) {
